Basic Yelp API Interaction and Routes
Adds a basic functionality to interact with the Yelp API, including
logging in and receiving the access token. This token is requested
once at application start and will not be requested again until
the applicaiton is restarted. The application will currently fetch
bars from Yelp and sort them by rank. Additionally added a few
dependencies for running the application, including request to make
API calls from the YelpAPI service easier.