Creates Location Based Angular Frontend
Uses the location service to send data between the main controller
and the results controller. Once the position is retrieved in the
location service, this position is passed to the main controller
which then sets the current location of the location service. The
service then broadcasts a new location event which causes the
results controller to fetch the data from the API.