Adds Location Searching in Frontend
Adds the ability to uilize the search bar to search for locations
without using the geolocation function. Shares location data
between the results and main controller through the location
service. A new event is broadcast when the location is set in the
location service to trigger the results controller to retrieve the
results. Also includes a div to show error messages which may be
encountered.