Basic Attendance Routing for Post Requests
Adds a new attendance controller and basic attendance model. The
attendance model holds username, place and date. Updates the routes
file to link to the attendanceController, which currently only has one
route. Posts to /attendance/new shall include a JWT in order to create
a new successful attendance entry, if the JWT does not exist or is
not valid, then an error message will be sent.