Prevents Users from Attending an Event Twice
Modifies the attendance controller to check if there is already an
attendance for the particular user on that day. If that is the case,
an HTTP 418 response code is sent. Modifies front end code to handle
users attempting to register to attend a bar twice. Installs
react-toastify to present an error message upon recieving an error,
this package also requires webpack css loaders to run during the
webpack build. Installs style-loader and css-loader packages and
updates the webpack.common.js to add these to development and
production webpack builds.