Adds Basic Backend Logout Functionality
Creates a new TokenBlacklist model to hold blacklisted tokens in a
database. Modifies the JWT generation to include a UUID to identify
each token (JWT - jti ). Implements a delete route to /auth/token and
reads the token from the authorization-token header. If this token is
valid and not already in the token blacklist database, then it will
insert the token id (JWT - jti) into the database. Modifies the
verify token service to require a callback since the database is
being queried.