Frontend Modifications to Canx Attendance
Adds a several new actions to support HTTP deletes to /attendance/:id
no change of state occurs upon a successful delete; however, a new
getAttendance action is dispatched following a successful delete. Adds
a new button to the Result component which is only visible if the
current user is logged in and registered to attend that venue. Also,
updates the CSS for the button group so that it is centered in the
card.